Fold CancelRules into cache makefile
Since the benefits of canceling pattern rules apply mainly to
larger projects that would benefit even more from minion_cache,
and since it requires generation of an included makefile, it
seems to make sense to kill both birds with one stone.

The only concern is that it introduces the potential for
a difference in behavior, since the makefile could be leveraging
a pattern rule (unintentionally?) which will stop working with
minion_cache.  But the README says we disable pattern rules...